# Improved series for the meridian length and its inverse