--- a/src/projections/gn_sinu.cpp
+++ b/src/projections/gn_sinu.cpp
@@ -38,7 +38,7 @@ static PJ_LP gn_sinu_e_inverse (PJ_XY xy, PJ *P) {          /* Ellipsoidal, inve
     PJ_LP lp = {0.0,0.0};
     double s;
 
-    lp.phi = pj_inv_mlfn(P->ctx, xy.y, P->es, static_cast<struct pj_opaque*>(P->opaque)->en);
+    lp.phi = pj_inv_mlfn(xy.y, static_cast<struct pj_opaque*>(P->opaque)->en);
     s = fabs(lp.phi);
     if (s < M_HALFPI) {
         s = sin(lp.phi);