This PR addresses #2387   point 3.

Replace the series for the meridian length in the functions
```c++
pj_enfn
pj_mlfn
pj_inv_mlfn
```
in `mlfn.cpp` with improved versions.  The old version of `pj_mlfn` used Snyder's Eq. (3-21) extended from 3rd to 4th order in e^2 and `pj_inv_mlfn` inverted this by an iterative method.

The new version uses 6th order expansions in the third flattening, n, for both `pj_mlfn` and `pj_inv_mlfn`.  Expansions in n have a smaller truncation error that the corresponding expansion in e^2.  Increasing the order to 6th (from 4th) means that the routines give full double precision accuracy for |f| <= 1/150.  Because the meridian length is used in several projections, it's important that it be accurate as possible.  For example, now the azimuthal equidistant projection is equally accurate in the polar (uses `pj_mlfn`) and oblique (uses the geodesic routines) aspects.

The increased accuracy tripped one of the tests: an instance of the projection `+proj=utm +approx`.  The fix is to use the non-approx results for the expected values.

Presumably this change results in `pj_mlfn` being a little slower while `pj_inv_mlfn` is considerably faster (since there's no iteration involved).  Incidentally the new method for `pj_inv_mlfn` uses Snyder's recommended series Eq. (3-23) extended from 4th to 6th order (Snyder writes n as e_1).

More discussion of the series used here is given in my recent paper

> On auxiliary latitudes, https://arxiv.org/abs/2212.05818

<!-- Feel free to remove check-list items aren't relevant to your change -->

 - [x] Partially closes #2387
 - [x] Tests adjusted
 - [x] Added clear title that can be used to generate release notes