@rouault 2022-12-17
Can you try running the new ``bench_proj_trans`` binary before and after your changes ?
I get for example the following with current master:
```
$ bin/bench_proj_trans --pipeline "+proj=pipeline +step +proj=unitconvert +xy_in=deg +xy_out=rad +step +proj=utm +zone=31 +ellps=GRS80 +approx" 2 49
2 49 -> 426857.98771659 5427937.52334353
Duration: 407 ms
Throughput: 12.29 million coordinates/s

$ bin/bench_proj_trans --pipeline "+proj=pipeline +step +inv +proj=utm +zone=31 +ellps=GRS80 +approx +step +proj=unitconvert +xy_in=rad +xy_out=deg" 426857.98771659 5427937.52334353
426857.98771659 5427937.52334353 -> 1.99999999999991 48.9999999999999
Duration: 714 ms
Throughput: 7.00 million coordinates/s
```