@cffk 2022-12-17
Here you go:
```
BEFORE
2 49 -> 426857.98771659 5427937.52334353
Duration: 305 ms
Throughput: 16.39 million coordinates/s
426857.98771659 5427937.52334353 -> 1.99999999999991 48.9999999999999
Duration: 546 ms
Throughput: 9.16 million coordinates/s

AFTER
2 49 -> 426857.98771659 5427937.52334229
Duration: 333 ms
Throughput: 15.02 million coordinates/s
426857.98771659 5427937.52334353 -> 1.99999999999969 49.000000000011
Duration: 496 ms
Throughput: 10.08 million coordinates/s
```

These timings are roughly in line with my expectations.